requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
PyMySQL==1.0.3
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
